Avoid polluting the Nightdive dir with Ironwail-specific user data
Previously, the Nightdive dir was used as a "userdir"
when playing the 2021 re-release version of the game
in order to have access to the installed add-ons, and
because the Steam game dir might not be writable.

This caused issues, though: if there's an `mg1` subdir
in the Nightdive dir, KEX Quake will no longer look for
data in rerelease/mg1, so DOTM is no longer playable.
Same thing happens with the other expansions that ship
with the base game (rogue/hipnotic/dopa).

To avoid this, we now use a separate dir when playing
the Steam/EGS/GOG version of the game:

- Windows:
    %UserProfile%\Ironwail\rerelease (remastered version)
    %UserProfile%\Ironwail\original (original version)

- Linux:
    ~/.ironwail/.rerelease (remastered)
    ~/.ironwail/.original (original)

Note: add-ons downloaded using the Mods menu are still
saved to the Nightdive dir (along with addons.json), so
that they can also be played in KEX Quake without having
to re-download anything.